math: log2f could generate spurious INEXACT exceptions
log2f(x) = logf(x) / logf(2.0f). The code used a double precision
constant for logf(2.0f) which appears to encourage gcc to generate an
INEXACT exception even when x is qNAN. Switch to a static const float
value of 1.0f / M_LN2 instead to ensure this is computed at compile
time and avoid a runtime exception.

I suspect there are more issues like this lurking in the math library,
but this is the only one I've found with the existing test suites.

Signed-off-by: Keith Packard <keithp@keithp.com>