[ARM] Fix FPSCR initial state for Arm8.1-M low overhead loops.
The Arm8.1-m architecture added Low Overhead Loops and tail predication,
controlled by the LTPSIZE bits of FPSCR. Unfortunately 0 is an invalid
value, with values other than 0x4 causing an LE instructions to throw a
UsageFault. This alters the initial reset value in _start to 0x40000 to
ensure it is initialized correctly.