tinystdio: fix compiling freopen on AVR
AVR puts functions in a non-default address space.
Unfortunately, Clang doesn't fully support this and crashes with the
following code:

    pf->read = (void *) read;
    pf->write = (void *) write;

However, this cast seems a bit odd to me: it casts from a function
pointer to a void* pointer and then back to a function pointer (to store
in the pf struct). Instead, this patch casts to the desired function
signature instad. This gets the code to compile for Clang/AVR.