Add more checks to test/tls.c
First, this adds an overaligned variable which ensures that the test
is likely to fail if ALIGN_WITH_INPUT is removed from the .tdata section
in the linker script. I am using this to allow testing linking with
ld.lld.

Second, we now check that the TLS data in RAM matches the data in ROM
(again this should always be true unless there is a bug in the linker or
linker script).

Finally, also check that the addresses of TLS variables are within the
TLS regions to verify that _set_tls() sets the value expected by the
compiler.

I originally made these changes to test linking with ld.lld instead of
ld.bfd, but it turns out they were also useful to expose the ARM
aligned TLS limitation fixed in the previous commit.