picolibc.ld.in: Don't use ALIGN_WITH_INPUT when linking with ld.lld
ld.lld uses a different LMA/VMA algorithm from ld.bfd and should always
ensure that the LMA/VMA padding between sections matches, so does not
require the ALIGN_WITH_INPUT attribute which was added to handle a
behaviour change in ld.bfd 2.23.2.
This patch removes the attribute as part of the configure_file() call when
linking with ld.lld. The alternative to this approach would be to use
a separate memcpy() for .tdata instead of relying on matching padding, but
that would require touching multiple places in the code and updating the
using.md documentation.

Checking the picolibc build-time linker is not perfect since in theory it
would be possible for some users to build picolibc with ld.lld and use
ld.bfd to link the final application. However, this case should be
handled by the newly added assertion that the flash and ram size match and
will result in a linker error. Building the tls.c test (now that it
contains an overaligned variable) with ld.bfd (using the ld.lld linker
script) results in the following errors:
```
/usr/bin/ld.bfd: ERROR: .data/.tdata flash size does not match RAM size
/usr/bin/ld.bfd: test/tls: section `.tdata' can't be allocated in segment 1
```