picolibc.ld.in: drop the i MEMORY attribute when linking with ld.lld
This attribute is not supported by ld.lld and does not appear to result
in any difference when linking with ld.bfd. I have compared binaries built
both with and without this attribute in the linker script and they are
completely identical.

Looking at the ld.bfd source code, it appears that the MEMORY section flags
are only used when detecting which region to use (lang_memory_default), but
since all sections have an associated "AT>", so we could actually drop this
unconditionally. I decided to keep this for now in case downstream projects
do rely on this flag for orphan sections.