math: Make function aliases for correct types
If float and double are the same size, then alias float functions
to the double names.

If double and long double are the same size, then alias double functions
to the long double names.

If double is not 64 bits, but long double is, then use the 'double' code for
long double instead.

This also gets rid of the separate files containing code for
_DOUBLE_IS_32BITS, making things smaller.

Signed-off-by: Keith Packard <keithp@keithp.com>