arm/aarch64: handle TLS alignment in _set_tls()
If the TLS region contains overaligned variables, we have to add padding
after the tcb region to ensure that it has the same alignment as the actual
TLS data. Without this change, the newly-extended tls.c test was failing
since the addresses of the TLS variables were incorrect since the
R_ARM_TLS_LE32 relocation evaluated to an offset of 0x80 (due to the 128
byte alignment) for the first TLS variable but _set_tp() was only
subtracting 8. I noticed this due to the ARM CI failures, but I also
changed AArch64 since they use the same TLS scheme (Variant I).

To obtain the TLS segment alignment I added a new __tls_align variable to
the linker script. This should be more efficient than looking at the PT_TLS
PHDR (which we currently can't accesses anyway).