avr: work around Clang crash
Clang has a bug where it will crash for some inputs when built with
assertions enabled. While this is a bug in Clang (and not in picolibc),
it does reveal a possible issue in picolibc. So I figured I'd send a
patch to picolibc as well as fixing it in Clang.

Fix to Clang: https://reviews.llvm.org/D138681