Provide an empty libg
The default configuration in GCC on many platforms adds -lg to the
linker command line if the compilation is using debug (-g).
glibc has provided an empty libg.a since at least 1995, newlib provides a
libg.a hardlinked to libc.a, and so on.
Having this empty library avoids linker errors when using a custom
toolchain with picolibc as the default libc and not using picolibc.specs

Signed-off-by: Joakim Nohlgård <joakim@nohlgard.se>