picolibc.ld: Reduce number of memory region attributes
Place all read-only or executable regions in flash, place all read-write
sections in RAM.

The memory region attributes in the ld script are used to select a
suitable region for orphan sections, or output sections where the
destination address or region has not been set explicitly.

If a section matches _ANY_ of the attributes listed before the '!' and
matches NONE of the attrbutes after '!' then it will be placed in that
region. Therefore it does not make sense to add too many attributes here
when we only have two regions to choose from. In particular, having the
'a' attribute listed for both regions will defeat the purpose of all the
other attributes because ld will pick the first one for all ALLOCATABLE
sections.

References:
 - https://sourceware.org/binutils/docs/ld/MEMORY.html
 - https://sourceware.org/binutils/docs/ld/Output-Section-Address.html
 - https://sourceware.org/binutils/docs/ld/Output-Section-LMA.html

Signed-off-by: Joakim Nohlgård <joakim@nohlgard.se>