Rewrite TLS initialization and allocation
The initial TLS template image is placed in flash and a separate area is
allocated for the TLS variables for the initial thread. crt0 calls
_init_tls to set up the initial TLS area. The calculated size of the
combined TLS area (.tdata + .tbss) needs to account for different
alignment requirements for .tbss and .tdata, this is done by aligning
the size of .tdata to the alignment of .tbss.
Add a function _tls_align() to get the alignment requirement for the TLS
area.
This change also eliminates the need for ALIGN_WITH_INPUT on the .tdata
section in the linker script because it is no longer meant to piggyback
on the .data initialization code in crt0.