picocrt/x86: Keep 32-bit stack 16-byte aligned for exit
Recent GCC versions use SSE registers for stack manipulation when
available and when building with speed optimization. This means we
need to be careful about stack alignment in assembly code, including
crt0-32.S. The call to main() is made with an empty stack, but the
call to exit passes a parameter. Pad the stack with 3 extra values to
ensure it remains aligned to 16-bytes.

Signed-off-by: Keith Packard <keithp@keithp.com>