Fix 'startfile' value in picolibc.specs
Gcc specs files are tricky -- the %s operator (which performs a
library file lookup) only operates within a parameter, not after the
parameter has been finished. This means that it must be applied directly to the
value and not after any subsequent {} operations.

Reformat the startfile expression to construct complete filenames and
apply %s to those, rather than attempting to construct them in pieces
with brace expressions building part of the name and then applying %s
afterwards.

Signed-off-by: Keith Packard <keithp@keithp.com>