Wrap the various crt0 objects into static libraries.
When linking an application against picolibc, if I want to choose
between its various choices of `crt0-foo.o` objects (say, `crt0.o` vs
`crt0-semihost.o`), the only way I know to do it is to put the
appropriate object file name on the linker command line. This requires
knowing the _absolute path_ of that file, which is unergonomic: it
involves automated build systems having to find out that path and
write it into makefiles, and worse, manually written makefiles might
have to hard-code the pathname and then not be able to run on a
machine where the toolchain is installed in a different directory.

It would be nicer to be able to write a short option that just says
'use `crt0-semihost.o`, please' and have the linker automatically look
in `$sysroot/lib` to find it.

The best answer I've found is to have each crt0 object be wrapped in a
tiny library file, so that instead of including a full path to
`crt0-foo.o` on the command line, I can just write `-lcrt0-foo`. I
don't know if this works in all linkers, but with lld at least it
seems fine: lld already knows it wants to find a symbol called _start,
and if that symbol isn't among its explicit object files, it will
search the libraries for it and find it in `libcrt0foo.a`.

This commit does not remove, or prevent installation of, the bare
`crt0.o` and friends. The new libraries are an alternative, but the
usual approach should still work too.