libc/signal: Make signal/raise only support C semantics
Switch signal array from per-thread to global, avoiding TLS overhead.
When available, use atomics (configurable) to access the array to make
it re-entrant.

To be consistent with POSIX, have raise reset the handler to SIG_DFL
if it was set to some value other than SIG_DFL or SIG_IGN. This is
explicitly allowed by C.

Signed-off-by: Keith Packard <keithp@keithp.com>