scripts: support Zawrs option from qemu version 8 for riscv
qemu added support for another risc-v option, Zawrs, and enabled it by
default. It requires the 'a' option though, so targets that aren't
using that option won't be able to support it.

Picolibc doesn't need the Zawrs option at all, so add it to the list
of available options but leave it disabled.

Signed-off-by: Keith Packard <keithp@keithp.com>