libc/posix: Always use unsigned char to index charjump array
The charjump array was originally indexed by a plain 'char' type; when
'char' was signed, the address was moved to the middle of the array so that
negative offsets would reach the correct locations.

That's really complicated and fragile; a few paths which wanted to
free the array weren't re-adjusting the address, and there was a mix
of signed and unsigned indexing going on.

Instead of switching everything back to signed chars, just change all
of the indexes to unsigned values. This has the advantage of being
more efficient on arm where chars are naturally unsigned but compilers
often make them signed for 'compatibility'.

Signed-off-by: Keith Packard <keithp@keithp.com>