libc/tinystdio: Fix building with -Dposix-io=false
Two new functions that take a mode string were added, freopen and
fmemopen. The former relies on POSIX I/O and so needed to only be
built when posix-io=true. The latter still needs a way to convert the
open mode string to the internal FILE operation flags. Do this by
creating a new API, __stdio_sflags, which sits atop __posix_sflags
when posix-io=true, but which directly uses a modified version of the
__posix_sflags implementation when posix-io=false.

Closes: #486.

Signed-off-by: Keith Packard <keithp@keithp.com>