picolibc.ld: Ensure .tdata alignment matches .tbss alignment
If no TLS block has been assigned yet, picolibc uses the template inside
the binary directly. However, ld.lld only guarantees that .tdata/.tbss
are valid as a TLS template and not necessarily directly usable when using
picolibc's existing initialization logic. For ld.lld, we add an explicit
alignment expression to the start of .tdata to ensure .tdata variables
are at least as aligned as the .tbss ones.
This fixes the tls test when linking with ld.lld while allowing to keep
the single memcpy() call.
For context on linker behaviour: https://github.com/picolibc/picolibc/pull/411

Also adds a test with .bss being more aligned than .tbss to ensure we
use the correct address for that.