libc/stdio: Fix legacy stdio positional arg handling
va_list might be declared as an array (as on x86_64), in which
case we must wrap the value inside a struct to support passing it
by address. Use va_copy to move it into the right place. This probably
messes up targets that do something in va_end.

On targets with sizeof(char *) > sizeof(int), the printf code that
attempts to stash intermediate values gets confused sometimes by *
arguments. Deal with this by just treating all 'int' args as 'char *'
in get_arg so that bits aren't lost. This assumes that these targets
align all arguments on char * instead of int boundaries.

Signed-off-by: Keith Packard <keithp@keithp.com>