test: Validate exception configuration for various float types
Targets with mixed hard/soft float implementations generally don't
support exceptions for types implemented in software. Because we want
to signal that support to applications in fenv.h, we need to know that
at compile time, not runtime. To validate that those settings are
correct, add a test to compare the compile-time values with runtime.

Signed-off-by: Keith Packard <keithp@keithp.com>