Cygwin: Fix Windows file handle leak in stat("file", -1)
Don't leak a Windows file handle if stat() is called with a valid
filename, but invalid stat buffer pointer.

We do not destroy fh (which closes a Windows handle it has opened) if an
exception happens in the __try block.

Avoid this by re-ordering things so that we don't construct the fhandler
object until after we've attempted to use the struct stat buffer.

Fixes: 73151c54d581 ("syscalls.cc (stat_worker): Don't call build_fh_pc with invalid pc.")
Signed-off-by: Jon Turney <jon.turney@dronecode.org.uk>